source/dashboard/requirements.txt
altair
streamlit
python-dotenv
boto3
aioboto3
asyncio
psycopg2-binary
bcrypt